# docs/openapi.yaml
openapi: 3.1.0
info:
  title: BlotGuard-AI API
  version: 1.0.0
servers:
  - url: /
security:
  - bearerAuth: []
paths:
  /api/v1/health/live:
    get:
      operationId: getLiveness
      security: []
      responses:
        "200":
          description: Process is alive
  /api/v1/health/ready:
    get:
      operationId: getReadiness
      security: []
      responses:
        "200":
          description: Service is ready
        "503":
          description: A dependency is not ready
  /api/v1/analyses:
    post:
      operationId: createAnalysis
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                localize:
                  type: boolean
                  default: false
      responses:
        "202":
          description: Analysis task accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisTask"
        "400":
          $ref: "#/components/responses/Error"
        "413":
          $ref: "#/components/responses/Error"
        "415":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/Error"
  /api/v1/analyses/{task_id}:
    parameters:
      - $ref: "#/components/parameters/TaskId"
    get:
      operationId: getAnalysis
      responses:
        "200":
          description: Analysis task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisTask"
        "404":
          $ref: "#/components/responses/Error"
    delete:
      operationId: deleteAnalysis
      responses:
        "204":
          description: Task deleted
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /api/v1/analyses/{task_id}/report:
    parameters:
      - $ref: "#/components/parameters/TaskId"
    get:
      operationId: downloadReport
      responses:
        "200":
          description: PDF report
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        "409":
          $ref: "#/components/responses/Error"
  /api/v1/artifacts/{artifact_id}:
    get:
      security:
        - bearerAuth: []
        - artifactCookie: []
      operationId: getArtifact
      parameters:
        - name: artifact_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: download
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Artifact binary
        "404":
          $ref: "#/components/responses/Error"
  /api/auth/login:
    post:
      operationId: login
      security: []
      description: Verify a database user and return a signed access token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Authenticated session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          $ref: "#/components/responses/FrontendError"
  /api/auth/register:
    post:
      operationId: register
      security: []
      description: Create a local user with a hashed password.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "201":
          description: Registered and authenticated session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          $ref: "#/components/responses/FrontendError"
        "409":
          $ref: "#/components/responses/FrontendError"
  /api/auth/logout:
    post:
      operationId: logout
      security: []
      responses:
        "204":
          description: Artifact cookie cleared; clients must discard their access token.
  /api/auth/me:
    get:
      operationId: currentUser
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current authenticated user
        "401":
          $ref: "#/components/responses/FrontendError"
  /api/tasks/upload:
    post:
      operationId: frontendUploadTask
      description: Frontend v0.1 compatibility upload endpoint.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "201":
          description: Frontend task summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FrontendTask"
        "400":
          $ref: "#/components/responses/FrontendError"
        "413":
          $ref: "#/components/responses/FrontendError"
        "415":
          $ref: "#/components/responses/FrontendError"
  /api/tasks/{task_id}:
    parameters:
      - $ref: "#/components/parameters/TaskId"
    get:
      operationId: frontendGetTask
      description: Frontend v0.1 compatibility polling endpoint.
      responses:
        "200":
          description: Frontend task summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FrontendTask"
        "404":
          $ref: "#/components/responses/FrontendError"
    delete:
      operationId: frontendDeleteTask
      description: Delete a terminal task, database rows, and stored files.
      responses:
        "204":
          description: Task deleted
        "404":
          $ref: "#/components/responses/FrontendError"
        "409":
          $ref: "#/components/responses/FrontendError"
  /api/tasks/{task_id}/result:
    parameters:
      - $ref: "#/components/parameters/TaskId"
    get:
      operationId: frontendGetTaskResult
      description: Frontend v0.1 compatibility result endpoint.
      responses:
        "200":
          description: Frontend result view
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FrontendResult"
        "404":
          $ref: "#/components/responses/FrontendError"
  /api/tasks/{task_id}/report:
    parameters:
      - $ref: "#/components/parameters/TaskId"
    get:
      operationId: frontendDownloadReport
      description: Frontend v0.1 compatibility PDF report download.
      responses:
        "200":
          description: PDF report
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        "404":
          $ref: "#/components/responses/FrontendError"
        "409":
          $ref: "#/components/responses/FrontendError"
components:
  securitySchemes:
    artifactCookie:
      type: apiKey
      in: cookie
      name: blotguard_artifact_token
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    TaskId:
      name: task_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    FrontendError:
      description: Frontend compatibility error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/FrontendError"
  schemas:
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    LoginResponse:
      type: object
      required: [access_token, token_type, expires_in, user]
      properties:
        access_token:
          type: string
        token_type:
          type: string
          const: Bearer
        expires_in:
          type: integer
          minimum: 1
        user:
          type: object
          required: [id, username, role]
          properties:
            id:
              type: integer
            username:
              type: string
            role:
              type: string
    FrontendTask:
      type: object
      required:
        - task_id
        - file_name
        - file_type
        - status
        - backend_status
        - progress
        - created_at
      properties:
        task_id:
          type: string
          format: uuid
        file_name:
          type: string
        file_type:
          type: string
          enum: [jpg, jpeg, jfif, png, tif, tiff, pdf, docx]
        file_size:
          type: [integer, "null"]
        status:
          enum: [pending, processing, completed, failed]
        backend_status:
          enum:
            - queued
            - extracting
            - inferencing
            - reporting
            - succeeded
            - failed
            - cancelled
        progress:
          type: integer
          minimum: 0
          maximum: 100
        created_at:
          type: [string, "null"]
          format: date-time
        completed_at:
          type: [string, "null"]
          format: date-time
        error_message:
          type: [string, "null"]
    FrontendResult:
      allOf:
        - $ref: "#/components/schemas/FrontendTask"
        - type: object
          required:
            - image_count
            - items
            - result_summary
            - mask_available
            - risk_level_is_experimental
            - is_mock
            - report_available
          properties:
            original_image_url:
              type: [string, "null"]
              description: Compatibility alias for the first result item.
            image_count:
              type: integer
              minimum: 0
            items:
              type: array
              items:
                $ref: "#/components/schemas/FrontendResultItem"
            result_summary:
              $ref: "#/components/schemas/AnalysisSummary"
            mask_available:
              type: boolean
            mask_image_url:
              type: [string, "null"]
              description: Compatibility alias for the first result item.
            localization_message:
              type: [string, "null"]
            overall_score:
              type: [number, "null"]
              minimum: 0
              maximum: 1
              deprecated: true
              description: Compatibility alias of score_generated.
            score_generated:
              type: [number, "null"]
              minimum: 0
              maximum: 1
              description: Canonical uncalibrated detector score.
            score_semantics:
              type: [string, "null"]
              const: uncalibrated_sigmoid_risk_score
            prediction:
              type: [string, "null"]
              enum: [generated, original, null]
            threshold:
              type: [number, "null"]
            overall_risk:
              type: [string, "null"]
              enum: [very_low, low, medium, high, very_high, null]
              deprecated: true
              description: Compatibility alias of risk_level.
            risk_level:
              type: [string, "null"]
              enum: [very_low, low, medium, high, very_high, null]
              description: Canonical experimental five-level risk.
            risk_level_label:
              type: [string, "null"]
            risk_level_semantics:
              type: string
              const: experimental_class_balanced_calibrated_risk
            risk_level_version:
              type: string
              const: experimental-platt-balanced-v1
            risk_level_is_experimental:
              type: boolean
              const: true
            suspect_regions:
              type: array
              items:
                type: object
              description: Empty until localization output is confirmed.
            model_probabilities:
              type: array
              items:
                type: object
              description: Empty until multi-generator probabilities are confirmed.
            model_version:
              type: [string, "null"]
            weight_sha256:
              type: [string, "null"]
            device:
              type: [string, "null"]
            is_mock:
              type: boolean
            report_available:
              type: boolean
            report_url:
              type: [string, "null"]
            processing_time:
              type: [number, "null"]
            conclusion:
              type: string
    FrontendResultItem:
      type: object
      required:
        - item_id
        - status
        - source_index
        - mask_available
        - risk_level_is_experimental
      properties:
        item_id:
          type: string
          format: uuid
        status:
          enum: [pending, succeeded, failed]
        source_name:
          type: string
        source_index:
          type: integer
          minimum: 1
        page_number:
          type: [integer, "null"]
          minimum: 1
        width:
          type: integer
        height:
          type: integer
        sha256:
          type: string
        prediction:
          type: [string, "null"]
          enum: [generated, original, null]
        score_generated:
          type: [number, "null"]
          minimum: 0
          maximum: 1
        score_semantics:
          type: [string, "null"]
          const: uncalibrated_sigmoid_risk_score
        threshold:
          type: [number, "null"]
        risk_level:
          type: [string, "null"]
          enum: [very_low, low, medium, high, very_high, null]
        risk_level_label:
          type: [string, "null"]
        risk_level_semantics:
          type: string
          const: experimental_class_balanced_calibrated_risk
        risk_level_version:
          type: string
          const: experimental-platt-balanced-v1
        risk_level_is_experimental:
          type: boolean
          const: true
        original_image_url:
          type: [string, "null"]
        mask_available:
          type: boolean
        mask_image_url:
          type: [string, "null"]
        mask_coverage:
          type: [number, "null"]
        localization_message:
          type: [string, "null"]
        error:
          type: [object, "null"]
    AnalysisTask:
      type: object
      required: [schema_version, task_id, status, input, items]
      properties:
        schema_version:
          const: "1.0"
        task_id:
          type: string
          format: uuid
        status:
          enum: [queued, extracting, inferencing, reporting, succeeded, failed, cancelled]
        input:
          type: object
        model:
          type: [object, "null"]
        summary:
          $ref: "#/components/schemas/AnalysisSummary"
        items:
          type: array
          items:
            type: object
        artifacts:
          type: array
          items:
            type: object
        warnings:
          type: array
          items:
            type: string
        error:
          type: [object, "null"]
        report_available:
          type: boolean
        report_url:
          type: [string, "null"]
    AnalysisSummary:
      type: object
      required:
        - total
        - succeeded
        - failed
        - generated
        - original
        - risk_level_is_experimental
      properties:
        total:
          type: integer
        succeeded:
          type: integer
        failed:
          type: integer
        generated:
          type: integer
        original:
          type: integer
        score_generated:
          type: [number, "null"]
          minimum: 0
          maximum: 1
          description: Maximum score among successful image results.
        score_semantics:
          type: [string, "null"]
          const: uncalibrated_sigmoid_risk_score
        prediction:
          type: [string, "null"]
          enum: [generated, original, null]
        risk_level:
          type: [string, "null"]
          enum: [very_low, low, medium, high, very_high, null]
        risk_level_semantics:
          type: string
          const: experimental_class_balanced_calibrated_risk
        risk_level_version:
          type: string
          const: experimental-platt-balanced-v1
        risk_level_is_experimental:
          type: boolean
          const: true
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
            request_id:
              type: [string, "null"]
    FrontendError:
      type: object
      required: [error, message, request_id]
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
        request_id:
          type: [string, "null"]